requirements.txt
streamlit
pandas
requests
python-dotenv
beautifulsoup4
selenium
pdfplumber
pillow
pytesseract
sentence-transformers
sqlmodel
psycopg2-binary
redis
celery
numpy